rename walletMock -> createMockSigner
