+++ packages/chain-ethereum/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./signatureData.js"
 export * from "./implementation.js"
-export * from "./walletMock.js"
+export * from "./createMockSigner.js"
