+++ packages/chain-solana/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./implementation.js"
-export * from "./walletMock.js"
+export * from "./createMockSigner.js"
